Pardo loops may contain optional pragmas.
The "pragma" is simply a string literal which is contained in a
pardo_pragma instruction immediately prior to the pardo loop it belongs to.

The intent is that the sip will choose a specialized pardo loop handler
depending on the text of the pragma.

Error checking to ensure that pragmas have legal and appropriate contents is left as a TODO